requirements.txt
pandas
numpy
scipy
statsmodels
google-generativeai>=0.8.0
python-dotenv
fastapi
uvicorn
gunicorn
python-multipart
requests
streamlit
